Update CurrencyData.properties to more recent version.
This is to fix the "time is more than 10 years from present" error due to
too old timestamps of currency info change. The recent data is taken from
http://hg.openjdk.java.net/jdk6/jdk6/jdk/file/ac1d168048bd/src/share/classes/java/util/CurrencyData.properties
but note that it's not a complete replace, some parts were dropped due to the
"too many other currencies" error happening otherwise (perhaps, a new version
of GenerateCurrencyData.java is also needed to make it work but that's too much
for now).